validation loss and early stopping

+++ ssar/utils.py
@@ -1,6 +1,7 @@
 import os
 import torch
 import glob
+import numpy as np
 
 # Freezes or unfreezes a model
 def dfs_freeze(model, unfreeze=False):
@@ -15,23 +16,26 @@ def get_date_sorted_chkpt_files(checkpoint_path):
     return old_saves
 
 # Save the model
-def save_model(model, optimizer, epoch, step, checkpoint_path):
+def save_model(model, optimizer, epoch, step, best_val_loss, checkpoint_path, filename_override=None):
     print('Saving model...')
 
     # Remove old saved models to preserve space
-    old_saves = get_date_sorted_chkpt_files(checkpoint_path)
-    while(len(old_saves) > 5):
-        try:
-            os.remove(old_saves[0])
-        except:
-            print(f'Warning: Could not remove old checkpoint file at {os.path.realpath(old_saves[0])}')
-        del old_saves[0]
+    if filename_override is None:
+        old_saves = get_date_sorted_chkpt_files(checkpoint_path)
+        while(len(old_saves) > 5):
+            try:
+                os.remove(old_saves[0])
+            except:
+                print(f'Warning: Could not remove old checkpoint file at {os.path.realpath(old_saves[0])}')
+            del old_saves[0]
 
     # Save model
-    save_file_path = os.path.join(checkpoint_path, f'ssar_save_{epoch}_{step}.pth')
+    filename = f'ssar_save_{epoch}_{step}.pth' if filename_override is None else filename_override
+    save_file_path = os.path.join(checkpoint_path, filename)
     states = {
         'epoch': epoch,
         'step': step,
+        'best_val_loss': best_val_loss,
         'arch': 'ssar',
         'state_dict': model.state_dict(),
         'optimizer': optimizer.state_dict(),
@@ -53,7 +57,11 @@ def load_latest(model, checkpoint_path, optimizer=None):
         model.load_state_dict(checkpoint['state_dict'])
         if optimizer is not None:
             optimizer.load_state_dict(checkpoint['optimizer'])
-        return checkpoint['epoch'], checkpoint['step']
+        if 'best_val_loss' in checkpoint:
+            best_val_loss = checkpoint['best_val_loss']
+        else:
+            best_val_loss = np.inf
+        return checkpoint['epoch'], checkpoint['step'], best_val_loss
     else:
         print("INFO: No existing checkpoint file to load")
-        return 0, 0
+        return 0, 0, np.inf
